[ASV-1603] Handling mopub consent dialog management on cobrand
