+++ app/src/cobrand/java/cm/aptoide/pt/FlavourApplicationModule.java
@@ -4,6 +4,9 @@ import cm.aptoide.accountmanager.AdultContent;
 import cm.aptoide.pt.abtesting.experiments.MoPubBannerAdExperiment;
 import cm.aptoide.pt.abtesting.experiments.MoPubInterstitialAdExperiment;
 import cm.aptoide.pt.abtesting.experiments.MoPubNativeAdExperiment;
+import cm.aptoide.pt.ads.MoPubConsentDialogManager;
+import cm.aptoide.pt.ads.MoPubConsentDialogView;
+import cm.aptoide.pt.ads.MoPubConsentManager;
 import cm.aptoide.pt.ads.WalletAdsOfferCardManager;
 import cm.aptoide.pt.ads.WalletAdsOfferManager;
 import cm.aptoide.pt.preferences.AdultContentManager;
@@ -54,5 +57,20 @@ import javax.inject.Singleton;
 
   @Singleton @Provides WalletAdsOfferCardManager providesWalletAdsOfferCardManager() {
     return new WalletAdsOfferCardManager();
+  }
+
+  @Singleton @Provides MoPubConsentManager providesMoPubConsentManager() {
+    return new MoPubConsentManager();
+  }
+
+  @Singleton @Provides @Named("mopub-consent-dialog-view")
+  MoPubConsentDialogView providesMoPubConsentDialogView(MoPubConsentManager moPubConsentManager) {
+    return moPubConsentManager;
+  }
+
+  @Singleton @Provides @Named("mopub-consent-dialog-manager")
+  MoPubConsentDialogManager providesMoPubConsentDialogManager(
+      MoPubConsentManager moPubConsentManager) {
+    return moPubConsentManager;
   }
 }
