+++ app/src/main/java/cm/aptoide/pt/ApplicationModule.java
@@ -76,8 +76,6 @@ import cm.aptoide.pt.ads.AdsRepository;
 import cm.aptoide.pt.ads.MinimalAdMapper;
 import cm.aptoide.pt.ads.MoPubAdsManager;
 import cm.aptoide.pt.ads.MoPubAnalytics;
-import cm.aptoide.pt.ads.MoPubConsentDialogManager;
-import cm.aptoide.pt.ads.MoPubConsentDialogView;
 import cm.aptoide.pt.ads.MoPubConsentManager;
 import cm.aptoide.pt.ads.PackageRepositoryVersionCodeProvider;
 import cm.aptoide.pt.ads.WalletAdsOfferCardManager;
@@ -265,7 +263,6 @@ import com.jakewharton.rxrelay.BehaviorRelay;
 import com.jakewharton.rxrelay.PublishRelay;
 import com.liulishuo.filedownloader.FileDownloader;
 import com.liulishuo.filedownloader.services.DownloadMgrInitialParams;
-import com.mopub.common.MoPub;
 import com.twitter.sdk.android.core.TwitterAuthConfig;
 import com.twitter.sdk.android.core.TwitterCore;
 import com.twitter.sdk.android.core.identity.TwitterAuthClient;
@@ -1177,21 +1174,6 @@ import static com.google.android.gms.auth.api.Auth.GOOGLE_SIGN_IN_API;
       MoPubConsentManager moPubConsentDialogManager) {
     return new MoPubAdsManager(moPubInterstitialAdExperiment, moPubBannerAdExperiment,
         moPubNativeAdExperiment, walletAdsOfferManager, moPubConsentDialogManager);
-  }
-
-  @Singleton @Provides MoPubConsentManager providesMoPubConsentManager() {
-    return new MoPubConsentManager(MoPub.getPersonalInformationManager());
-  }
-
-  @Singleton @Provides @Named("mopub-consent-dialog-view")
-  MoPubConsentDialogView providesMoPubConsentDialogView(MoPubConsentManager moPubConsentManager) {
-    return moPubConsentManager;
-  }
-
-  @Singleton @Provides @Named("mopub-consent-dialog-manager")
-  MoPubConsentDialogManager providesMoPubConsentDialogManager(
-      MoPubConsentManager moPubConsentManager) {
-    return moPubConsentManager;
   }
 
   @Singleton @Provides Retrofit providesSearchSuggestionsRetrofit(
